Change record key to date (truncate in DAY)
The previous record key was CountryCode, and it didn't make sense, since
the key must be the same in order to group records (and the required
grouping is per day). It was truncated per day, so the minutes from
the original records and the minutes from the dummy record will be the
same, which guarantees the window closing.